+++ maps-sdk/src/test/java/com/mapbox/maps/shadows/ShadowSettingsService.kt
@@ -1,0 +1,7 @@
+package com.mapbox.maps.shadows
+
+import com.mapbox.common.SettingsService
+import org.robolectric.annotation.Implements
+
+@Implements(SettingsService::class)
+public class ShadowSettingsService