+++ maps-sdk/src/test/java/com/mapbox/maps/shadows/ShadowSettingsServiceFactory.kt
@@ -1,0 +1,7 @@
+package com.mapbox.maps.shadows
+
+import com.mapbox.common.SettingsServiceFactory
+import org.robolectric.annotation.Implements
+
+@Implements(SettingsServiceFactory::class)
+public class ShadowSettingsServiceFactory